+++ src/components/atoms/index.ts
@@ -4,5 +4,3 @@ export { default as Fab } from './Fab'
 export { default as TextInput } from './TextInput'
 export { default as TextBox } from './TextBox'
 export { default as Thumbnail } from './Thumbnail'
-export { default as MaterialDesignCard } from '../atoms/MaterialDesignCard'
-export { default as FlatDesignCard } from '../atoms/FlatDesignCard'
