+++ src/components/organisms/MaterialDesignCard.tsx
@@ -1,15 +1,22 @@
 import React from 'react'
 import { View, Text, Image, StyleSheet, ImageSourcePropType } from 'react-native'
 import { colors } from '../../themes'
+import { formatedDate } from '../../services/formatedDate'
+import { Button } from '../organisms'
+import { NavigationStackProp } from 'react-navigation-stack'
 
 type Props = {
   uri: ImageSourcePropType
   name: string
-  date: string
+  date: Date
   width: number
+  navigation: NavigationStackProp
+  onPressDetail: () => void
+  onPressApply: () => void
 }
 
-const FlatDesignCard: React.FC<Props> = props => {
+const MaterialDesignCard: React.FC<Props> = props => {
+  const date = formatedDate(props.date)
   return (
     <View>
       <View style={styles.imageBorderRadius}>
@@ -18,9 +25,27 @@ const FlatDesignCard: React.FC<Props> = props => {
       <View style={styles.description}>
         <View>
           <Text style={styles.name}>{props.name}</Text>
-          <Text style={styles.date}>{props.date}</Text>
+          <Text style={styles.date}>{date}</Text>
         </View>
-        {props.children}
+        <View style={styles.buttonContainer}>
+          <View style={styles.buttonWrapper}>
+            <Button
+              color={'#FFFFFF'}
+              fullWidth={false}
+              width={70}
+              height={30}
+              padding={6}
+              onPress={props.onPressDetail}
+            >
+              <Text style={styles.buttonText}>詳細</Text>
+            </Button>
+          </View>
+          <View style={styles.buttonWrapper}>
+            <Button color={'#FFFFFF'} fullWidth={false} width={70} height={30} padding={6} onPress={props.onPressApply}>
+              <Text style={styles.buttonText}>参加</Text>
+            </Button>
+          </View>
+        </View>
       </View>
     </View>
   )
@@ -28,10 +53,6 @@ const FlatDesignCard: React.FC<Props> = props => {
 
 const styles = StyleSheet.create({
   imageBorderRadius: {
-    borderBottomLeftRadius: 0,
-    borderBottomRightRadius: 0,
-    borderTopRightRadius: 16,
-    borderTopLeftRadius: 16,
     overflow: 'hidden'
   },
   image: {
@@ -53,10 +74,6 @@ const styles = StyleSheet.create({
     justifyContent: 'space-between',
     padding: 6,
     flexDirection: 'row',
-    borderBottomLeftRadius: 16,
-    borderBottomRightRadius: 16,
-    borderTopRightRadius: 0,
-    borderTopLeftRadius: 0,
     overflow: 'hidden'
   },
   buttonContainer: {
@@ -72,4 +89,4 @@ const styles = StyleSheet.create({
   }
 })
 
-export default FlatDesignCard
+export default MaterialDesignCard
