+++ src/components/organisms/index.ts
@@ -6,3 +6,5 @@ export { default as UserListItem } from './UserListItem'
 export { default as ApplyModal } from './ApplyModal'
 export { default as Button } from './Button'
 export { default as Card } from './Card'
+export { default as MaterialDesignCard } from './MaterialDesignCard'
+export { default as FlatDesignCard } from './FlatDesignCard'
